Move calibration code to the calibration flavor source folders
