+++ app/src/calibration/java/org/covidwatch/android/di/FlavorSpecificModules.kt
@@ -1,6 +1,11 @@
 package org.covidwatch.android.di
 
+import org.covidwatch.android.domain.ExportDiagnosisKeysAsFileUseCase
+import org.covidwatch.android.domain.ProvideDiagnosisKeysFromFileUseCase
+import org.covidwatch.android.exposurenotification.KeyFileSigner
+import org.covidwatch.android.exposurenotification.KeyFileWriter
 import org.covidwatch.android.ui.reporting.NotifyOthersViewModel
+import org.koin.android.ext.koin.androidApplication
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -11,6 +16,26 @@ val flavorSpecificModule = module {
             exportDiagnosisKeysAsFileUseCase = get(),
             enManager = get(),
             positiveDiagnosisRepository = get()
+        )
+    }
+
+    single { KeyFileSigner() }
+    single { KeyFileWriter(androidApplication(), get()) }
+
+    factory {
+        ProvideDiagnosisKeysFromFileUseCase(
+            enManager = get(),
+            diagnosisKeysTokenRepository = get(),
+            contentResolver = androidApplication().contentResolver,
+            dispatchers = get()
+        )
+    }
+
+    factory {
+        ExportDiagnosisKeysAsFileUseCase(
+            enManager = get(),
+            keyFileWriter = get(),
+            dispatchers = get()
         )
     }
 }
