+++ app/src/calibration/java/org/covidwatch/android/ui/MainActivity.kt
@@ -4,25 +4,19 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
 import android.os.Parcelable
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.lifecycleScope
-import org.covidwatch.android.R
 import org.covidwatch.android.domain.ProvideDiagnosisKeysFromFileUseCase
 import org.covidwatch.android.domain.ProvideDiagnosisKeysFromFileUseCase.Params
 import org.covidwatch.android.extension.launchUseCase
 import org.koin.android.ext.android.inject
 import timber.log.Timber
 
-
-class MainActivity : AppCompatActivity() {
+class MainActivity : BaseMainActivity() {
 
     private val provideDiagnosisKeysFromFileUseCase: ProvideDiagnosisKeysFromFileUseCase by inject()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
         handleIntent(intent)
     }
 
@@ -38,17 +32,5 @@ class MainActivity : AppCompatActivity() {
                 lifecycleScope.launchUseCase(provideDiagnosisKeysFromFileUseCase, Params(it))
             }
         }
-    }
-
-    private val FragmentManager.currentNavigationFragment: Fragment?
-        get() = primaryNavigationFragment?.childFragmentManager?.fragments?.first()
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        supportFragmentManager.currentNavigationFragment?.onActivityResult(
-            requestCode,
-            resultCode,
-            data
-        )
     }
 }
