+++ app/src/main/java/org/covidwatch/android/di/AppModule.kt
@@ -25,8 +25,7 @@ import org.covidwatch.android.data.pref.PreferenceStorage
 import org.covidwatch.android.data.pref.SharedPreferenceStorage
 import org.covidwatch.android.domain.*
 import org.covidwatch.android.exposurenotification.ExposureNotificationManager
-import org.covidwatch.android.exposurenotification.KeyFileSigner
-import org.covidwatch.android.exposurenotification.KeyFileWriter
+
 import org.covidwatch.android.ui.Notifications
 import org.covidwatch.android.ui.exposurenotification.ExposureNotificationViewModel
 import org.covidwatch.android.ui.exposures.ExposuresViewModel
@@ -54,9 +53,7 @@ val appModule = module {
         )
     }
 
-    single { KeyFileSigner() }
-    single { KeyFileWriter(androidApplication(), get()) }
-    single { SafetyNet.getClient(androidApplication()) }
+        single { SafetyNet.getClient(androidApplication()) }
 
     single {
         SafetyNetManager(
@@ -155,15 +152,6 @@ val appModule = module {
     }
 
     factory {
-        ProvideDiagnosisKeysFromFileUseCase(
-            enManager = get(),
-            diagnosisKeysTokenRepository = get(),
-            contentResolver = androidApplication().contentResolver,
-            dispatchers = get()
-        )
-    }
-
-    factory {
         UploadDiagnosisKeysUseCase(
             enManager = get(),
             diagnosisRepository = get(),
@@ -184,13 +172,7 @@ val appModule = module {
         )
     }
 
-    factory {
-        ExportDiagnosisKeysAsFileUseCase(
-            enManager = get(),
-            keyFileWriter = get(),
-            dispatchers = get()
-        )
-    }
+
 
     factory {
         UpdateExposureStateUseCase(
