+++ app/src/main/java/org/covidwatch/android/ui/BaseMainActivity.kt
@@ -1,0 +1,28 @@
+package org.covidwatch.android.ui
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import org.covidwatch.android.R
+
+open class BaseMainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+
+    private val FragmentManager.currentNavigationFragment: Fragment?
+        get() = primaryNavigationFragment?.childFragmentManager?.fragments?.first()
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        supportFragmentManager.currentNavigationFragment?.onActivityResult(
+            requestCode,
+            resultCode,
+            data
+        )
+    }
+}
