+++ app/src/production/java/org/covidwatch/android/di/FlavorSpecificModules.kt
@@ -1,5 +1,7 @@
 package org.covidwatch.android.di
 
+import org.covidwatch.android.ui.reporting.NotifyOthersViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val flavorSpecificModule = module {
