+++ app/src/production/java/org/covidwatch/android/ui/MainActivity.kt
@@ -1,0 +1,4 @@
+package org.covidwatch.android.ui
+
+
+class MainActivity : BaseMainActivity()
