+++ app/src/production/java/org/covidwatch/android/ui/menu/Menu.kt
@@ -1,8 +1,5 @@
 package org.covidwatch.android.ui.menu
 
-import org.covidwatch.android.R
-import org.covidwatch.android.data.exposureinformation.ExposureInformationRepository
-
 class MenuFragment : BaseMenuFragment()
 
-class MenuAdapter(onClick: (destination: Destination) -> Unit) : BaseMenuAdapter(onClick)
+class MenuAdapter(onClick: (menuItem: MenuItem) -> Unit) : BaseMenuAdapter(onClick)
